requirements.txt
psycopg2-binary==2.9.6
tabulate==0.9.0
PyQt5~=5.15.9
Flask==2.3.2
requests==2.31.0